# packages/backend/openapi.yaml
openapi: 3.0.3
info:
  title: TrustAnchor Engine 1 — Verification Server API
  version: 1.0.0
  description: |
    This server is storage and serving ONLY (Frozen Architecture Specification §5).
    It holds no signing key of any kind, in any environment, at any time. Every
    signature that matters was produced offline by @trustanchor/offline-signer;
    this server simply stores and serves the resulting data.

    Read GET /revocation's description carefully before using it — it is a
    convenience, unsigned, derived view and is NOT part of the actual
    cryptographic verification algorithm (which reads revocation status
    directly from the already-fetched, already-signature-verified GET
    /manifest response).
servers:
  - url: http://localhost:4000
    description: Local development
paths:
...
  /asset/{contentHash}:
    get:
      summary: Fetch a raw asset by its SHA-256 content hash
      description: |
        NOT called by Engine 1 (Frozen Spec §13) — reserved for a future
        Engine 2 (out of scope for this implementation) to download a
        trustworthy reference image for visual comparison.
      parameters:
        - name: contentHash
          in: path
          required: true
          schema: { type: string, pattern: '^[0-9a-f]{64}$' }
      responses:
        '200':
          description: Raw asset bytes.
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        '400':
          description: contentHash is not a well-formed 64-character hex SHA-256 digest.
        '404':
          description: No asset found for this content hash.
    post:
      summary: Ingest a raw asset file (administrative)
      security:
        - ingestionAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
                docId: { type: string, format: uuid }
                assetName: { type: string, example: student_photo }
                expectedContentHash: { type: string, pattern: '^[0-9a-f]{64}$' }
      responses:
        '201':
          description: Asset ingested successfully.
        '400':
          description: Uploaded bytes' hash does not match expectedContentHash.
...
components:
  securitySchemes:
    ingestionAuth:
      type: http
      scheme: bearer
      description: |
        Ordinary operational access control (Frozen Spec §21), NOT a
        cryptographic security boundary. See src/middleware/ingestionAuth.ts.
...